translation_service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
transformers==4.30.0
torch==2.0.0
google-cloud-translate==3.11.0
